# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/tags": {
            "get": {
                "description": "List all tags",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tags"
                ],
                "summary": "List tags",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Include bookmark count for each tag",
                        "name": "with_bookmark_count",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter tags by bookmark ID",
                        "name": "bookmark_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search tags by name",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.TagDTO"
                            }
                        }
                    },
                    "403": {
                        "description": "Authentication required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "description": "Create a new tag",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tags"
                ],
                "summary": "Create tag",
                "parameters": [
                    {
                        "description": "Tag data",
                        "name": "tag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TagDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.TagDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "403": {
                        "description": "Authentication required"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.TagDTO": {
            "type": "object",
            "properties": {
                "bookmark_count": {
                    "description": "Number of bookmarks with this tag",
                    "type": "integer"
                },
                "deleted": {
                    "description": "Marks when a tag is deleted from a bookmark",
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}